create file for each class
