+++ person.rb
@@ -1,10 +1,6 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError
-  end
-end
+require './nameable.rb'
 
-class Person
+class Person < Nameable
   attr_reader :id
   attr_accessor :name, :age, :parent_permission
 
@@ -27,30 +23,5 @@ class Person
 
   def of_age?
     @age >= 18
-  end
-end
-
-class Decorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name()
-    @nameable.correct_name.upcase
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name[0, 10]
   end
 end
